fix: 푸드트럭 리뷰 평점 int -> float
With: S02P31C211-111-review-TDD

+++ Backend/src/main/java/com/boyworld/carrot/api/controller/review/response/FoodTruckReviewResponse.java
@@ -10,7 +10,7 @@ public class FoodTruckReviewResponse {
 
     private List<FoodTruckReviewDto> foodTruckReviewDtoList;
 
-    private int averageGrade;
+    private float averageGrade;
 
     @Builder
     private FoodTruckReviewResponse(List<FoodTruckReviewDto> foodTruckReviewDtoList){
